Create TOML V2
Given that the new things we're adding to the user file aren't
compatible with older versions of mantid, we're bumping up to TOML V2 so
that we can keep being backwards compatible. This is mainly to stop new
user files containing this information from being loaded by older
versions, making it clear that you need to upgrade to use these new
features and toml files.

RE #38524
