+++ scripts/SANS/sans/user_file/toml_parsers/toml_parser.py
@@ -8,6 +8,7 @@ from sans.state.AllStates import AllStates
 from sans.state.IStateParser import IStateParser
 from sans.user_file.toml_parsers.toml_reader import TomlReader
 from sans.user_file.toml_parsers.toml_v1_parser import TomlV1Parser
+from sans.user_file.toml_parsers.toml_v2_parser import TomlV2Parser
 
 
 class TomlParser(object):
@@ -33,5 +34,7 @@ class TomlParser(object):
         version = toml_dict["toml_file_version"]
         if version == 1:
             return TomlV1Parser(toml_dict, file_information=file_information)
+        if version == 2:
+            return TomlV2Parser(toml_dict, file_information=file_information)
         else:
             raise NotImplementedError("Version {0} of the SANS Toml Format is not supported".format(version))
