+++ scripts/SANS/sans/user_file/toml_parsers/toml_v1_schema.py
@@ -4,46 +4,19 @@
 #     NScD Oak Ridge National Laboratory, European Spallation Source
 #     & Institut Laue - Langevin
 # SPDX - License - Identifier: GPL - 3.0 +
-import re
+
+from sans.user_file.toml_parsers.toml_base_schema import TomlSchemaValidator
 
 
-class TomlValidationError(Exception):
-    # A key error would be more appropriate, but there is special
-    # handling on KeyError which escapes any new lines making it un-readable
-    pass
-
-
-class TomlSchemaV1Validator(object):
+class TomlSchemaV1Validator(TomlSchemaValidator):
     # As of the current TOML release there is no way to validate a schema so
     # we must provide an implementation
 
-    # Note : To future devs, if we have a V2 schema a lot of this class could
-    # be split into a SchemaValidator and an inheriting V1 and V2 schema
-    # would override the reference schema with the new one
-
     def __init__(self, dict_to_validate):
-        self._expected_list = self._build_nested_keys(self._reference_schema())
-        self._to_validate_list = self._build_nested_keys(dict_to_validate)
-
-    def validate(self):
-        self._to_validate_list = filter(lambda s: not s.startswith("metadata"), self._to_validate_list)
-        unrecognised = set(self._to_validate_list).difference(self._expected_list)
-
-        if not unrecognised:
-            return
-
-        # Build any with wildcards
-        wildcard_matchers = [re.compile(s) for s in self._expected_list if "*" in s]
-        # Remove anything which matches any the regex wildcards
-        unrecognised = [s for s in unrecognised if not any(wild_matcher.match(s) for wild_matcher in wildcard_matchers)]
-
-        if len(unrecognised) > 0:
-            err = "The following keys were not recognised:\n"
-            err += "".join("{0} \n".format(k) for k in unrecognised)
-            raise TomlValidationError(err)
+        super(TomlSchemaV1Validator, self).__init__(dict_to_validate)
 
     @staticmethod
-    def _reference_schema():
+    def reference_schema():
         """
         Returns a dictionary layout of all supported keys
         :return: Dictionary containing all keys, and values set to None
@@ -137,32 +110,6 @@ class TomlSchemaV1Validator(object):
             },
         }
 
-        component_keys = {
-            "idf_component_name": None,
-            "device_name": None,
-            "device_type": None,
-            "location": {"x", "y", "z"},
-            "transmission": None,
-            "efficiency": None,
-            "empty_cell": None,
-            "initial_polarization": None,
-        }
-        filter_keys = dict(component_keys, **{"cell_length": None, "gas_pressure": None})
-        field_keys = {
-            "sample_strength_log": None,
-            "sample_direction": {"a", "p", "d"},
-            "sample_direction_log": None,
-        }
-        polarization_keys = {
-            "flipper_configuration": None,
-            "spin_configuration": None,
-            "flipper": {"*": component_keys},
-            "polarizer": filter_keys,
-            "analyzer": filter_keys,
-            "magnetic_field": field_keys,
-            "electric_field": field_keys,
-        }
-
         return {
             "toml_file_version": None,
             "binning": binning_keys,
@@ -175,28 +122,4 @@ class TomlSchemaV1Validator(object):
             "q_resolution": q_resolution_keys,
             "reduction": reduction_keys,
             "transmission": transmission_keys,
-            "polarization": polarization_keys,
         }
-
-    @staticmethod
-    def _build_nested_keys(d, path="", current_out=None):
-        if not current_out:
-            current_out = []
-
-        def make_path(current_path, new_key):
-            return current_path + "." + new_key if current_path else new_key
-
-        for key, v in d.items():
-            new_path = make_path(path, key)
-            if isinstance(v, dict):
-                # Recurse into dict
-                current_out = TomlSchemaV1Validator._build_nested_keys(v, new_path, current_out)
-            elif isinstance(v, set):
-                # Pack all in from the set of names
-                for name in v:
-                    current_out.append(make_path(new_path, name))
-            else:
-                # This means its a value type with nothing special, so keep name
-                current_out.append(new_path)
-
-        return current_out
