+++ scripts/test/SANS/user_file/toml_parsers/CMakeLists.txt
@@ -1,4 +1,6 @@
-set(TEST_PY_FILES toml_parser_test.py toml_v1_parser_test.py toml_v1_schema_test.py)
+set(TEST_PY_FILES toml_parser_test.py toml_v1_parser_test.py toml_v1_schema_test.py toml_v2_parser_test.py
+                  toml_v2_schema_test.py
+)
 
 add_subdirectory(parser_helpers)
